requirements.txt
# Core scientific stack
pandas>=2.0
numpy>=1.26
scipy>=1.11
matplotlib>=3.7

# HTTP + scraping + PDF
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
pymupdf>=1.24
pdfplumber>=0.10
tqdm>=4.66

# Bibliometrics + topic modeling
sentence-transformers>=2.7
umap-learn>=0.5.5
scikit-learn>=1.4

# Graph analysis
networkx>=3.2
python-louvain>=0.16

# Geo + metadata
pycountry>=24.0

# Scholar scraping (optional)
scholarly>=1.7

# Document generation
python-docx>=1.0
openpyxl>=3.1
